Move Bean Validation Test to Spring MVC
