+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/web/application/SpringBootMvcApplicationService.java
@@ -19,6 +19,6 @@ public class SpringBootMvcApplicationService {
   }
 
   public JHipsterModule buildUndertowModule(JHipsterModuleProperties properties) {
-    return factory.buildUntertowModule(properties);
+    return factory.buildUndertowModule(properties);
   }
 }
