+++ src/main/resources/generator/server/springboot/mvc/web/test/BeanValidationAssertions.java.mustache
