+++ src/main/resources/generator/server/springboot/mvc/web/test/BeanValidationTest.java.mustache
